Provisioning via template. Now having issues adding secret.
